fixed wall collisions and camera locations
Wall collisions adjusted so that I no longer have to past ground objects on top of walls to stop the player from clipping through.

Improved camera physics such that it remains locked to a 45 degree angle viewpoint unless mouse entry, upon which it will apply mouse velocity to x and y coordinates. Needs work to move to more applicable locations with less effort but overall works somewhat well and avoids the dreaded "camera.look_at(player)" pitching bug by including the "return_rotation" variable.

Steps forward - work on texturing the base ground in order to include the 5 row strategy, adjust W and S inputs to jump the player exactly one space to the left or right in order to facilitate the 5-lane movement. DESPERATELY NEEDS A DOSE OF OOP.

Recommended to consider for assignment project proposal as this is actually quite interesting and has the potential to fuflil a lot of the project criteria while also including the ideas I wanted to add. Also think of a name better than 3D DASH, it sounds like an awful geometry dash ripoff (which it is, but I'd prefer not to give people that impression before they start playing).