Movement and Collision Update
Updated Z-axis movement to operate off a 5-lane switching system, where W moves up one lane and S moves down one. Reaching the end of the lane will block inputs that could potentially move the player character over the edge of the collision. I might want to try animating a red flash and shake if the player tries to go off the edge.

Also updated wall and ground definitions to operate off an Object-oriented class based system to allow for faster entry and assignment to coordinate values. More conversion to OOP and optimisation principles is recommended if I decide to choose this as my project.

Next step is to implement somewhat better camera motion to show the speed of the player character, as well as test out adding renders of levels as part of the ground collisions through blender or obj files.