Slight improvements
Added inability for the player to simply jump off the edge, restricting them to five lanes.
Adjusted the spawn height as it was causing issues with the gravity and making the player clip through the ground on accident.
Made the camera follow the player while maintaining all normal motion of returning to an equilibrium position, rather than being locked to the starting zone. variable return_location is now global and constantly updating to the player's exact location as the game progresses, ensuring that the return location is not fixed.

Later improvements to make - test jump heights as well as blender integration to level design - I would rather not progressively build each level in a seperate file and import them, it would be much nicer to have a single ground/wall entity that instakills on wall touch but maintains normal playability on ground touch. Maybe consider separate-entity spikes or other assorted obstacles that are imported seperately in order to have directionless instakill touches?