Floor collision
I FINALLY DID IT AAAAAAAAAAAAAAAAAAAAAAA Floor collision is technically functional, the player can move on the lowest level, but wall collisions instantly crash the game and any elevations do not work and clip the player into the ground (possibly due to the average of possible heights).

possibly try using preset collision with an adjusted editor camera? not sure.