FINALLY
boxcastcollider.py contains the best collision capabilities - able to interpret collision from any .obj map imported as well as other vertices mapped in-line, without using any form of preset.

Current system can be shown by enabling debug on the lower body boxcast collider. Essentially draws an invisible cube that detects collisions from the playercharacter. collision logic contained in " if hit_info.hit" statement.

Next steps for working - the current collision system detects by setting the player position to the highest point that the box collider hits. Thus, if the player jumps into a wall, it will act as if the top part of the wall doesn't exist, and the player will float inside of the wall at the same y value that it initially touched the wall. Possible resolutions for the "walls will kill you" collider, simply add a secondary box collider at the top half of the player model (to avoid accidental floor collisions) and use to detect anything in front of the player in an incredibly short range. This could be implemented by forming a boxcast at the player's xyz coordinates but limiting the scale to x/2, y/2, z (where each coordinate value is derived from the playermodel, and the boxcollider extends inside the player and not forwards from it, thus causing it to only trigger once the player actively intersects with a wall.

Also, redesign the level meshes or at least increase the X scale, the player seems to always overshoot jumps and perspective is a little off. (possibly tweak camera too, the current system is still hard to properly use for distance calculations.