Explosions and more
Started developing functionality for death animation, cube now explodes on hit with collision. Has massive framerate issues on lower end devices, so I'd suggest limiting the cell fractures on the explosion to around 25 rather than 100.

Maybe work on a jump animation similar to the cube flip in geometry dash. Be sure to save the blender files in a non-baked format such that it can be properly changed for skin adjustments. (possibly not needed, texture applications might be more time-efficient on the ursina-end, if so then animations need to be output as white only to make a proper canvas.