cleaning up
Cleaned up excess files and other components that weren't necessary to functionality. Also improved fps rate for lower end devices by lowering the cell fractures on the player model's explosion. Final resolution issue is to make sure the fps is capped, as higher end devices override the attempted 60fps cap, and the playerspeed is dependent on the cap. One of those two situations needs to be resolved.