Movement and Controls Update
Decoupled the movement of the player from the internal fps counter, the game now theoretically runs at the maximum possible v-sync frames (automatically calculated by Ursina based on hardware settings), but the internal logic operates on a 60-tick framework, where logic such as collisions, movement speed, and game updates are handled in sets of 60 ticks per second, thus making the system more approachable to lower end devices, more capable of predictability, less susceptible to being seriously impacted by random frame drops, and more consistent between high-end and low-end laptops, both for development time and in eventual release.

Furthermore, I have updated the control scheme to be more intuitive - Q/A testing the latest version showed that players frequently attempted to switch lanes with A and D rather than W and S, and claimed the controls were misleading as that was the movement directions they were used to. Based on feedback, I have temporarily adjusted the movement controls to be A for switching one lane to the left and D for one lane to the right. Furthermore, I have added controls for an automatic player reset (r), a fullscreen mode toggle (f), and an instant quit (esc). Menus and pause controls will be initiated by pressing TAB once implemented.

Finally, I've restructured the code architecture to be more organised and intuitive. I will not adjust for OOP components just yet, but that will be happening soon.

Next steps are to design the menu layout, most likely through the usage of a storyboard for diagramming. I will also setup final tests for the hitbox controls, as there are still a few issues in development causing occasional bugs that seem difficult to resolve at the current time.