Menu System Update
Added extremely basic support for game menus and controls. The game now preloads assets with a visible loading screen before putting the player on the main menu. All proposed menu buttons are active with assigned functions, but do not have any purpose assigned to them besides a redirect and a "WIP" text prompt. The only functional system currently is the level select, and even then, it's only a prototype with a single selectable level, that being the preset level I designed at the beginning of the project for testing collision. I have also implemented a rudimentary pause menu (it doesn't even have a quit button, just resume), that pauses the game upon pressing tab.

Next steps are to work on level hotswapping from the menu, adding functionality to the pause menu, and working on the databases for level data and player data. Would be best to test on PC so I can work on making a few more simplistic levels for testing. Will also give me time to determine whether the game works cross-platform properly now.

Customisation options shouldn't be too hard to implement so I'll leave that on the backburner for now.