New Levels and Knowledge
After developing a second demo level for the game, I have discovered new ideas to test and things to understand.

- BDSF colours applied through nodes are only recognisable to blender. Colours must be applied normally for ursina rendering

- The scale modifier for ursina is dependent on total model height, as are the coordinate values, not a single point. Thus, I need to find a new way to align the model's base section with the player spawn section.

- I need both the mtl and obj files for object import handling. mtl handles individual colours and is called automatically by renderers when accessing the obj file. Keep them both together, and don't namechange without re-exporting from blender or adjusting the file routing in the obj file (and vice versa for mtl)

- Always, and I mean always, check that the model has been exported with the correct settings. I need to check off 'export with triangulated mesh' in order for ursina to interpret anything.

The levels are interchangeable as expected, adjusting the value for the safeground mesh import (need to change the name of that, it's still artifacting from previous versions) will thus adjust what level I import. In later versions, I need to change that file's location and set it to different components based on menu changes, or at least cycle through a list and insert it to a generalised level initialiser once locked in.

I've also added an integrity test file. When exporting to pc, almost everything worked, except that my first change unintentionally destabilised Ursina's ability to locate the files for the explosion death animation. I reverted the file through github and it did fix the change, but just in case, I've written a file to check the program's ability to locate the required animation files. It's in it's own python file for now, but can be quickly added or imported to the main file and used.

Next steps are to work on level select options and finally get started on actually building levels for once.