Positioning and Tracking
The player's percentage position throughout the movement of a level is now automatically confirmed. Should have been a simple job but due to blender import issues, the model needs to be rotated 270 degrees before drawing, so attempting to reason vertex locations without first applying the transform causes issues. This was encountered when I first tried, where Ursina's rotation of the GameMap entity didn't inherently rotate the initial vertex positions that the script drew from. In order to resolve this, I had to apply trigonometry (Who said it was useless? I need to have a polite discussion with them about how often it saves me) in order to calculate the proper location of the coordinate rotations.

This worked initially, until I realised that my blender export scheme was inherently incompatible; my original plans were to essentially export a level with 20m in front of (0, 0, 0) and everything else behind, however the trigonometric calculation required that the length of track before and after the (0, 0, 0) world point was equal. Thus, I had to update the models for the 2nd level to shift them closer to the centre of the blender area and therefore the centre of the game.

Currently, the progress percentage export is just a print statement of percentage; next steps are to fill in the placeholder function for creating the progress bar and percentage indicator.

Also there was a hilarious mess-up where I placed the calcpoints function within the update loop, so instead of running it once at the start of the level, it would run continuously every frame, causing so much damage that my pc, previously running the window at over 300fps, suddenly dropped to 4 and then crashed to desktop within twenty seconds.