More Level Specificity and Separation
This update involves the development of inter-level systems that allow for a higher degree of separation between the levels.

In this, I have developed a system for seamless level additions and subtractions through OOP inheritance principles. All level components are identified through the .json file and the names of each of the level .obj files. By inserting a "Levelx.obj" file into the system and a "Levelx" : "0" parameter into the .json file (where x is any given level number), it will thus incorporate the new elements seamlessly into the system with no extra file setup.

Furthermore, the .json file logs progress into the levels, only updating if the progress percentage is higher than before, and loads to .json after each death (restarting not included).

The level select screen also updates the background colour as an RGB scale range identified from the HSV colourmap that I stole from matplotlib. The values for a given level are converted to a fractional scale from 0 to 1 and assigned a randomised color from that.

As a result of these changes, the game can now theoretically support any number of levels as long as an obj file and level name is added as proposed.

Further expansion from this point is to actually add an 'end' to the level to identify the 100% range to allow people to complete the level fully. I should also look at pre-baked lighting rendering to determine whether I can make a given level more readable.