Ending and Final Saving Update
During level progress, a counter at the top of the screen and a progress bar now show level percentage completion up to one decimal place. This completion is stored within the level data json file which also links to the main levels list by updating map components.

Levels now have an 'ending' by spawning a large green wall at the end of the map. This is only a prototype wall for the time but functions as the end of the level by colliding with it. Colliding with the wall triggers an 'endscreen' that allows the player to return to the main menu and saves the score as 100.0%

Things to patch and update next:
- Add music in some way; or at least sound effects.

- Update winscreen to be less of a randomly harsh cut to the screen, maybe an audio stinger and a fade-in or shake explosion effect (similar to geometry dash finish). Something to add a feeling of accomplishment.

- ADD MORE LEVELS. There are only two and they are quite short. Some form of tutorial may be in order to add too.

- Finally get around to player customisation and other similar additions. Actually add an options menu for goodness sake.

- Menu revamp is another option but I'll leave that for the afterburner because there are more important things to handle first.