Zeroday patch and font upgrade
Updated the fonts to now be actually interesting and look more thematically appropriate to the boxy theme

Patched the zero-day flaw for constant options menu entitycramming, it now stores and deletes the entities when appropriate.

Next system to develop is the customisation menu system and the skins associated. Window component saving should also be done.