Customisation Menu Progress and Shading
Updated the customisation menu to now properly render and remove entities when necessary. I have left the debug lines in as proof of concept for this commit, but they will be removed in the next commit for clarity.

Furthermore, I have allowed the cube to be maneuverable by mouse input, similar to the camera movements of the main gameplay loop.

I've started on attempts to run shaders as well, they don't seem to have too much of a performance impact, at least on a mid-low range pc, but I will test on my laptop as well to confirm for significantly lower end users. Currently shaders only seem to work on ursina-rendered entities, as the level meshes don't have any rendered shaders despite my own efforts.

I'll also be continuing the work I had in this for the wireframe models, as I think they add much-needed depth to the models to help with player visualisation of the 3D environment. In a similar vein, I will try to add a slight camera shift delay on jumping such that the player doesn't seem rooted to the camera.

And finally, of course, I will try to adjust menu buttons to actually control skins for the player, as currently it just makes a nice affirming statement saying I know how to use inheritance principles, and I'd prefer it to actually do something.

Next commit date is unconfirmed, could be a few hours, a day or a week depending on time and length of update.