Customisation Menu Completion (Nearly)
I have now fully finished and unified the menu systems for the customisation segment. The user can now select a variety of skins from a series of buttons with the colour or skin shown. A dropdown menu provides a message for the name of the skin, whether that be the color definition or the name of the image texture. The player's choice is saved to the skindata file and displayed on a reference cube in the centre of the screen. The player can also interact with this cube to rotate it on different axes and see different sides. This will be especially helpful once I work out full texturemapping for objects, currently I'm just pasting the same image on all six faces.

Still to be done:

- The locking system for allowing me to set up certain skins to need to be unlocked through progress. The integrated system is present, I just need to update the json file when levels are completed, and integrate the  prewritten locking mechanism for disabling skin access to a premade lock icon and tooltip display.

- Furthermore, I need to finish uploading a few more skins, currently the one without a tooltip is being replaced by a backup skin I have specifically designed for skins that are loaded faulty, and in this case the skin name is blank due to non existence. This will be updated soon.

Other things to finish - updating the audio to.. actually have audio, developing a few more levels, and considering the addition of animations. Shader effects are also a consideration, though their presence is largely extremely frustrating and I am not sure whether I will be able to complete it properly.