Locks, Deaths and a Whole Lot More Assets
Completed tasks:
Integrated the lock system to autonomously update as I change the number of levels and skins available. You may notice in the level_data.json file that level3 and level4 exist and have preset values, that was for testing level components to see whether skins would unlock unconditionally, which I have since patched. As long as the number of levels aligns with the number of locked skins (where I have placed each 'lockable' skin to button index position 13 and above), unlocked skins will be updated to the level completions after each successful level completion. Data is also stored afterwards, so skins are unlocked fully after a single completion of it's associated level.

Updated the main death animation to properly integrate skins. All explosion animations are now doubled in entity count in order to allow the wireframe model to have a death animation that doesn't conflict with the others. This also makes the other skins have more intricate and defined model appearances that help differentiation from the environment.

I also 3D modelled and edited a lock icon for the lock system, I think it looks way more visually appealing than a simple grey box and a tooltop, especially for smaller devices or people who have a harder time reading the text (as well as language barriers).

I also patched a few bugs in camera positioning and player respawn mechanisms which would occasionally result in tracking and positioning errors, or even making the game appear as if it had frozen while simultaneously locking the camera, hiding the player, and enabling movements. This has since been patched by me.

Next steps:
Level 3 and 4 are now MANDATORY to complete in order to fulfil the remaining two skins and make them accessible. Furthermore, I need to revamp Level 1 and make the current level 2 become the new level 1, as it is significantly easier, as determined from from A/B testing results.

Additionally, I should consider audio for the levels, whether as simple as SFX on user input or more complicated music tracks.

Furthermore, I need to keep working to patch the level shading issue, as levels become more complex, the ability for players to discern individual components and routes will reduce significantly. Thus, this issue must be rectified immediately.

Finally, and it is still as optional as ever, adding more animations to the player would be fun. I'm not sure though, it seems like it could cause issues in rendering and lower-end device capabilities.