Massive Textures and Sounds Update
Added sounds. That's right, I added sounds. It really wasn't too hard. What was actually difficult was getting the sounds. While the player's laneswitch sound effect is stolen from the hit 2008 action film "Jumper" and the button click was ripped from a free asset store, the background music for the main menu was entirely composed by myself using Audacity for pitch shifting, tempo changes and compositing. I will consider making more music for the background of the levels, but it will need to remain a lower background music style to avoid intrusive sounds, rather than something actually synced to the track.

Textures:
I've now added new textures for a loading screen and menu screen that doesn't look like it was built in 2012 HTML code. It now featues a custom synthwave style cylinder loop that repeats itself after the first movement for as long as the player is within the tube. Model was once again constructed by me, I have the blender files for proof if necessary. I've also built an addendum to the loading screen to have actually interesting animations rather than just staring at the word loading for a minute. As more animations are added, the loading screen will only take longer, so to spice it up I added a percentage bar and a silly little cube that bounces around to keep your attention.

I've also patched a small bug in the options menu that, if the player were to enter a very specific sequence of button presses, would cause the pause menu to render over the main menu when exiting options. For reference, this sequence was (Open game -> level select -> play level -> pause -> open options -> close options -> quit to menu -> open options -> close options). Despite this, my own testing and from rehired playtesters (begging my family for help), it was found to be a common thing to occur, and thus needed to be patched.

The next things to work on for this project are to continue adding sounds, develop the final music pieces, and make the last two levels. I also need to design a skybox, and my current idea is to roll with the synthwave theme and use the model I made for the menu animations as a custom skybox. I'm not sure what Ursina needs for skyboxes but if all else fails I can just import it really massive and pretend it's just a skybox. I might consider adding a few levels past that, as well as a few more achievement skins to match, but that's entirely a matter of available time and energy.