Options Menu Update, skyboxes, and a few Patches
Added new functionality to the options menu. Players can now rebind their input keys to any button on their keyboard or click on their mouse. This is set for the left and right dashes, as well as the jump. I mean very much any button, and tests from hired (begged) testers have shown that apparently windows, alt, shift, enter, ctrl, and even the function keys can be used as inputs. I might consider stripping those away, but honestly if someone decides they want that to be their input key, it is by no means my fault.

Skyboxes - I'm attempting to start on making the skybox controls, and it is currently only able to show images. I can't use models due to the requirement for the Sky() function to be an image. Thus, I need to either define a model around the playable area or figure out how to apply shifting images to the skybox.

Patch:
Moved skinapply to the top of the main function area in order to allow for the player's model in the loading screen to properly update to their selected preferences.

Next steps are to keep working on the skybox and make more audio additions, as well as some more levels.