Skyboxes, Levels and Failures
Added a skybox system for cylinder generation in the approximate style of the menu. This is generated within Ursina rather than loading a model, as constant errors have shown that the model conflicts with other rendered entities and is really difficult to resource-manage.

Level - added a level 3, a variant of level 1 with extra components added and improvements to existing level components. This is intended to be a very difficult level for players to do, and one to be learned over time. Difficulty will likely continue ramping.

Failure - attempted to add a gravity modifier, but extreme time loss over multiple hours yielded very little results and showed that the framework for my gravity and collision system is difficult to invert. This will either require a multi-commit time investment or an idea to be considered another time, as it is nonessential to gameplay. I have left the gravflipsfx.mp3 file added in case I decide to try again.

Next steps: More levels, skins and audio to be added until I decide otherwise. The game is currently in an extremely playable state. I should, however, consider optimisations to the 250 image files for the main menu screen, as they comprise about 800mb alone and I'd prefer to simplify that for lower end devices, or at least add a menu toggle to avoid rendering so much data.