I don't think you understand the gravity of the situation (UPDATE)
Added gravity swapping! I didn't think it was possible after my last failure but the answer was, naturally, working smarter and not harder. I was massively overcomplicating the system and drawing way too many classes when in most cases I could simply rely on the logic that gravity is a constant value that's either positive or negative, and use that to dictate how different systems should function rather than calling entirely new classes. The gravity system is now fully operational, at least for switching from downwards to upwards. I'll need to design a new level to test upwards to downwards gravity transitions, as currently the system only supports one gravity change cube per level and I need to update the internal systems to support multiple. This will require a rewrite of most essential systems, so I'm using this upload as a checkpoint in order to ensure I can backdate in case things go wrong. Also, all of the gravity changes are in the gravtest.py file, so that will be the de-facto file from now on. Boxcastcollider.py has been permanently moved to the 'old stuff' folder, which will be deleted for the final version, along with any deprecated files and systems. I may also change the file name in the future to 'Main' or  something similar to reduce confusion.