Final Level Addition
With this commit, I have finally completed the last level in the main game. I may create more at another time, but for the purposes of level and skin demonstration, this will do.

I have also created a demo of a different camera system. The current one is my personal favourite, and the demo is still quite buggy, but I thought I'd do it just to prove that a spherical orbit camera system is possible instead of a movement based location system.

I've also added the gravityswap gates for level 4, I may consider adding more just to mess with players, but I currently feel that it is a good amount for the level, and I don't want to overload players with things to process.

Final steps for game completion are to reorganise file system and consider musical additions to levels. After that, I am finished.