Data Reorganisation Stage 1
Part 1 of multiple in reorganising files to cleaner positions and setting proper folder management. All files are now reallocated and their individual components are set to proper file systems. Most are already integrated into the game but I need to continue working on the main system for skin selection as it is currently nonfunctional and returning type-errors. TBD.

Also, I had the random idea to allow people to upload skins and I might try and do that as a side project for fun. Current plans for, if implemented, are to allow a file upload after all levels are 100% completed.