Bugfix Patch #1
Resolved the errors in the texture loading system, allowing for proper file pathing. Turns out that Ursina only works properly in texturing when using relative file paths, and breaks itself when trying to use direct ones. How frustrating.

Patched an audio rendering issue where components would occasionally not be loaded. Audio and volume loading is now located within the prerendering system to allow for more precise loading and ensure proper location handling.

Updated a flaw in the customisation menu that would cause an extra wireframe border to generate each time it was opened, ultimately adding unnecessary entities

Updated the skin selection buttons to use quads instead of cubes, allowing for better optimisation and menu load times

Patched a bug where the camera would occasionally not snap to the player after deaths under a very specific set of circumstances.

Resolved an error in the OOB detection system that would cause a similar issue as the one above.

Patched an asset loading error that would cause some components to not be prerendered properly. Turns out that Ursina dislikes using multithreading and occasionally has file overlap errors. This has, however, resulted in an error wherein the loading screen does not display anything, which will need to be addressed in the next bugfix.