Bugfix Patch #2
Reassigned all internal values to a new file - Main.py

Changed the options menu to return to it's old form. I have simply elected to ignore any errors in the file load processing as whatever is missed will inevitably be fixed by lazy loading, so no textures will have errors.

Added a 'reset all data' function to the main menu so that players can, if they so desire, reset all of their data. This is non-repairable if selected, and it also helps me to reset data for finalising the project submission.

Fixed the camera location on level start by setting a starting position and rotation in the renderMap function. This means that an old error of the camera lerp return accidentally setting to a randomly distant value is theoretically patched, however I will need to continue testing to ensure that the patch works properly.

Last remaining tasks to complete:
Update main blender model to show depth innately, floors and ceilings will have a darker green tone than platforms, and the sides of red walls will be darker than the fronts. This will be extremely beneficial to player depth perception. Suggestion credit to beta tester Simon who had the idea for a prebaked shader depth rather than an actively rendered one. While a fully prebaked depth shader would be nice, it is currently more efficient to simply adjust necessary depth components.

Also, consider the addition of a 'potato pc' mode that skips prerendering and disables the menu animation and player death animation that causes so much lag, as well as removing shader effects. This will be toggleable in the menu, but swapping from potato to normal will require a reloading textures menu.

Neither of these adjustments are necessary to game functionality and could be removed if necessary, but I think them important enough to consider for the final bugfix patch.