Model Update and Completion
And there we go! The game is in a completed state and all functions are active. I've tweaked the models to provide more visual depth and clarity and the debug statements are removed. Error warnings are still present for player clarity, however. There is also no way to disable the ursina debug settings and startup settings so I just have to hope that everything works fine in the final section for .exe output.

Enjoy the game, and good luck! <3