Emergency Hotfix
Well.. this wasn't supposed to happen as early as it did but apparently my potato mode had a few more bugs I didn't expect. It is now patched. Previous issues in pause menu resetting during a death animation combined with it to cause a series of entity destruction failures, pause menu resetting errors and a bunch of other annoying and frustrating things that shouldn't have happened. All are fixed now. Hopefully this is the last hotfix I'll have to release.